+++ app/Livewire/Settings.php
@@ -11,7 +11,7 @@ class Settings extends Component
     use \Livewire\WithFileUploads;
 
     public $setting_id = 0, $businessName, $phone, $taxpayerId, $vat, $printerName, $website, $leyend, $creditDays = 15, $address, $city, $creditPurchaseDays, $confirmationCode, $decimals;
-    public $checkStockReservation;
+    public $checkStockReservation, $salesViewMode;
     public $globalCommission1Threshold, $globalCommission1Percentage, $globalCommission2Threshold, $globalCommission2Percentage;
     public $globalAllowCredit, $globalCreditDays, $globalCreditLimit, $globalUsdPaymentDiscount;
     public $discountRules = [];
@@ -89,6 +89,7 @@ class Settings extends Component
             $this->globalCommission2Percentage = $config->global_commission_2_percentage;
             $this->logo_preview = $config->logo; // Load existing logo
             $this->checkStockReservation = (bool) $config->check_stock_reservation;
+            $this->salesViewMode = $config->sales_view_mode;
             $this->defaultWarehouseId = $config->default_warehouse_id;
             
             // Network Printer
@@ -258,6 +259,7 @@ class Settings extends Component
                 'global_commission_2_threshold' => $this->globalCommission2Threshold,
                 'global_commission_2_percentage' => $this->globalCommission2Percentage,
                 'check_stock_reservation' => $this->checkStockReservation ? 1 : 0,
+                'sales_view_mode' => $this->salesViewMode,
                 'default_warehouse_id' => $this->defaultWarehouseId,
                 'backup_emails' => $backupEmailsArray,
                 'purchasing_calculation_mode' => $this->purchasingCalculationMode,