+++ resources/views/livewire/pos/sales.blade.php
@@ -96,22 +96,20 @@
 
                         <div class="d-flex justify-content-between align-items-center mb-2">
                             <label class="mb-0"></label>
-                            <div class="custom-control custom-switch" title="{{ $sellerConfig ? '' : 'Seleccione un cliente con conf. de vendedor para habilitar' }}">
-                                <input type="checkbox" class="custom-control-input" id="switchApplyFreight" wire:model.live="applyFreight" {{ $sellerConfig ? '' : 'disabled' }}>
+                            <div class="custom-control custom-switch">
+                                <input type="checkbox" class="custom-control-input" id="switchApplyFreight" wire:model.live="applyFreight">
                                 <label class="custom-control-label" for="switchApplyFreight" style="font-size: 0.8rem;">
                                     Aplicar Solo Flete
-                                     @if(!$sellerConfig) <i class="fas fa-lock text-muted" style="font-size: 0.7em;"></i> @endif
                                 </label>
                             </div>
                         </div>
 
                         <div class="d-flex justify-content-between align-items-center mb-2">
                             <label class="mb-0"></label>
-                            <div class="custom-control custom-switch" title="{{ $sellerConfig ? '' : 'Seleccione un cliente con conf. de vendedor para habilitar' }}">
-                                <input type="checkbox" class="custom-control-input" id="switchFreight" wire:model.live="is_freight_broken_down" {{ $sellerConfig ? '' : 'disabled' }}>
-                                <label class="custom-control-label" for="switchFreight" style="font-size: 0.8rem;">
+                            <div class="custom-control custom-switch">
+                                <input type="checkbox" class="custom-control-input" id="switchBreakdownFreight" wire:model.live="is_freight_broken_down">
+                                <label class="custom-control-label" for="switchBreakdownFreight" style="font-size: 0.8rem;">
                                     Desglosar Flete
-                                    @if(!$sellerConfig) <i class="fas fa-lock text-muted" style="font-size: 0.7em;"></i> @endif
                                 </label>
                             </div>
                         </div>